Los params de los plugins son unsigned int
En versiones modernas de g++ y clang++ daba problemas tener un enum  y luego hacer cast a signed int

+++ HTTPInputPlugin/HTTPInputPlugin.cpp
@@ -464,11 +464,11 @@ const std::string HTTPInputPlugin::g_properties[] = {
 	"keyConfig"
 };
 
-const int HTTPInputPlugin::g_paramTypes[] = {
-	(int)(PARAM_ARRAY | PARAM_INPUT)
+const unsigned int HTTPInputPlugin::g_paramTypes[] = {
+	(PARAM_ARRAY | PARAM_INPUT)
 };
 
-const int * HTTPInputPlugin::getPropertiesType() const
+const unsigned int * HTTPInputPlugin::getPropertiesType() const
 {
 	return HTTPInputPlugin::g_paramTypes;
 }
